requirements.txt
alembic==1.17.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
Markdown==3.9
MarkupSafe==3.0.3
Pygments==2.19.2
python-dotenv==1.0.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
